\name{quantml}
\alias{quantml}
\alias{optEdges}
\alias{optNNI}
\title{Phylogenetic inference from quantitative traits using likelihood}
\usage{
	quantml(tree, Data)
	optEdges(object, ...)
	optNNI(object, ...)
}
\arguments{
	\item{tree}{phylogeny as an object of class \code{"phylo"}.}
	\item{Data}{data matrix with traits in columns.}
	\item{object}{object of class \code{"quantml"}.}
	\item{...}{optional arguments}
}
\description{
	To be completed.
}
\details{
	To be completed.
}
\value{
	An object (optimized or unoptimized, depending on the method of class \code{"quantml"}.
}
\references{
	Felsenstein, J. (1973) Maximum-likelihood estimation of evolutionary trees from continuous characters. \emph{American Journal of Human Genetics}, \bold{25}, 471-492.
	
	Felsenstein, J. (1981) Evolutionary trees from gene frequencies and quantitative characters: Finding the maximum likelihood estimates. \emph{Evolution}, \bold{35}, 1229-1242.
	
}
\author{Liam Revell \email{liam.revell@umb.edu}, Joan Maspons}
\seealso{
	
}
\examples{

}
\keyword{phylogenetics}
\keyword{maximum likelihood}
\keyword{phylogeny inference}
\keyword{statistics}
